Reftr Cand Prtl code + See Another User Cand Prtl
